refactor: Change model package to dto
